refact(modularization/nodes): moving nodes feature to modules
