+++ dashboard/src/features/settings/services/certificate.query.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import { fetch } from "@marzneshin/utils";
-import { NodesQueryFetchKey } from "@marzneshin/features/nodes";
+import { NodesQueryFetchKey } from "@marzneshin/modules/nodes";
 
 export async function fetchCertificate(): Promise<string> {
     return fetch('/nodes/settings').then((settings) => {
