+++ dashboard/src/features/subscription-settings/services/subscription-settings.mutate.ts
@@ -1,4 +1,4 @@
-import { NodeType } from "@marzneshin/features/nodes";
+import { NodeType } from "@marzneshin/modules/nodes";
 import { useMutation } from "@tanstack/react-query";
 import { fetch, queryClient } from "@marzneshin/utils";
 import { toast } from "sonner";
