+++ dashboard/src/modules/nodes/dialogs/settings/index.tsx
