+++ dashboard/src/modules/nodes/dialogs/settings/node-config-editor.tsx
