+++ dashboard/src/modules/nodes/log/container.tsx
