+++ dashboard/src/modules/nodes/services/create-node.mutate.ts
@@ -1,22 +1,18 @@
-
-import { NodeType, NodesQueryFetchKey } from "@marzneshin/features/nodes";
+import { NodeType, NodesQueryFetchKey } from "@marzneshin/modules/nodes";
 import { useMutation } from "@tanstack/react-query";
 import { fetch, queryClient } from "@marzneshin/utils";
 import { toast } from "sonner";
 import i18n from "@marzneshin/features/i18n";
 
-export async function fetchDeleteNode(node: NodeType): Promise<NodeType> {
-    return fetch(`/nodes/${node.id}`, { method: 'delete' }).then((node) => {
+export async function fetchCreateNode(node: NodeType): Promise<NodeType> {
+    return fetch('/nodes', { method: 'post', body: node }).then((node) => {
         return node;
     });
 }
 
-const NodesDeleteFetchKey = "nodes";
-
-
 const handleError = (error: Error, value: NodeType) => {
     toast.error(
-        i18n.t('events.delete.error', { name: value.name }),
+        i18n.t('events.create.error', { name: value.name }),
         {
             description: error.message
         })
@@ -24,17 +20,20 @@ const handleError = (error: Error, value: NodeType) => {
 
 const handleSuccess = (value: NodeType) => {
     toast.success(
-        i18n.t('events.delete.success.title', { name: value.name }),
+        i18n.t('events.create.success.title', { name: value.name }),
         {
-            description: i18n.t('events.delete.success.desc')
+            description: i18n.t('events.create.success.desc')
         })
     queryClient.invalidateQueries({ queryKey: [NodesQueryFetchKey] })
 }
 
-export const useNodesDeletionMutation = () => {
+
+const NodesCreateFetchKey = "nodes";
+
+export const useNodesCreationMutation = () => {
     return useMutation({
-        mutationKey: [NodesDeleteFetchKey],
-        mutationFn: fetchDeleteNode,
+        mutationKey: [NodesCreateFetchKey],
+        mutationFn: fetchCreateNode,
         onError: handleError,
         onSuccess: handleSuccess,
     })
