+++ dashboard/src/modules/nodes/services/logs.socket.ts
@@ -1,4 +1,4 @@
-import { NodeType } from "@marzneshin/features/nodes";
+import { NodeType } from "@marzneshin/modules/nodes";
 import useWebSocket, { ReadyState } from "react-use-websocket";
 import { useAuth } from "@marzneshin/features/auth";
 import { joinPaths } from "@tanstack/react-router";
