+++ dashboard/src/modules/nodes/tables/detail-table/index.tsx
