+++ dashboard/src/modules/nodes/types/index.ts
