+++ dashboard/src/routes/_dashboard/nodes/$nodeId/index.tsx
@@ -5,7 +5,7 @@ import {
 import {
     useRouterNodeContext,
     NodesSettingsDialog,
-} from "@marzneshin/features/nodes";
+} from "@marzneshin/modules/nodes";
 import { useDialog } from "@marzneshin/hooks";
 
 const NodeOpen = () => {
